Really stop using firewalld
On some centos7 builds there is no firewalld so we have to be a bit more
smarter about how we disable it. New method is to run an exec that stops
the service if it is running then use a package resource to uninstall it
completely. All of this happens before we install the iptables service
so they should not confict with each other.

Change-Id: I0750de9e75b63190531a3d39a5fcbb19f8e8c49e