Add support for openSUSE
openSUSE Leap uses its own firewall manager called SuSEfirewall2, which
is capable of loading custom iptables rules. This patch adds the
necessary configuration to tell SuSEfirewall2 where to look for custom
firewall rules so that we can manage openSUSE firewall rules in the same
way we manage firewall rules for other images. This is part of an effort
to add openSUSE images to nodepool.

Unlike the CentOS case, where we stop the firewalld service so that we
can manage rules directly with iptables, we don't disable SuSEfirewall2.
This is because on SUSE iptables is not a service so we can't manage it
in the same way we do on other operating systems. Alternatively, we
could add an iptables-restore to /etc/init.d/boot.local which would
bring up the rules at boot but going through SusEfirewall2 is less
divergent from ther other OS's and allows for refreshing the rules on a
running system if desired.

Change-Id: I135de72c423281202fe9d70e24aed8e20e2db264