Ensure IP tables is always running
After the first time puppet is applied using this module,
IP Tables will not be running. However, after its first
reboot, the service will start and the rules will be in effect.

Regardless, ensure the IP Tables service is running
and the rules applied during initial bringup of a node.
This will help avoid possible security incidents in case
a reboot is ever initiated.

Change-Id: Ie1079b38675fbd16f43df3afe173fcb1a07416de