Stop using connection tracking for ssh connections
There is an issue with newer kernels (it seems to happen with 4.15.x)
that when conntrack is reloaded while a connection has packets in
flight, this connection going forward is neither considered INVALID
nor RELATED nor ESTABLISHED by the stateful tracking. While this is
certainly a bug elsewhere, we can be nicer with the rest by not using
stateful tracking for ssh connections.

Change-Id: I5a82f0a0b7ddc1a58c12233fae5761569012eb11